+++ lib/ui/settings/SettingsPersonalInfoContentWidget.dart
@@ -26,6 +26,7 @@ import 'package:illinois/ui/onboarding2/Onboarding2ResearchQuestionnaireAcknowle
 import 'package:illinois/ui/onboarding2/Onboarding2ResearchQuestionnairePanel.dart';
 import 'package:illinois/ui/settings/SettingsWidgets.dart';
 import 'package:illinois/ui/widgets/RibbonButton.dart';
+import 'package:illinois/ui/widgets/SoundRecorderDialog.dart';
 import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/model/auth2.dart';
 import 'package:illinois/service/Analytics.dart';
@@ -38,6 +39,8 @@ import 'package:rokwire_plugin/ui/panels/modal_image_holder.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
+
+import '../../service/Storage.dart';
 
 class SettingsPersonalInfoContentWidget extends StatefulWidget {
   final String? parentRouteName;
@@ -137,6 +140,7 @@ class _SettingsPersonalInfoContentWidgetState extends State<SettingsPersonalInfo
       _PersonalInfoEntry(
           title: Localization().getStringEx('panel.profile_info.full_name.title', 'Full Name'),
           value: Auth2().account?.authType?.uiucUser?.fullName ?? ""),
+      _buildNamePronouncement(),
       _PersonalInfoEntry(
           title: Localization().getStringEx('panel.profile_info.middle_name.title', 'Middle Name'),
           value: Auth2().account?.authType?.uiucUser?.middleName ?? ""),
@@ -837,6 +841,7 @@ class _SettingsPersonalInfoContentWidgetState extends State<SettingsPersonalInfo
     );
   }
 
+
   void _onTapDeleteData() async {
     final String groupsSwitchTitle = Localization().getStringEx('panel.settings.privacy_center.delete_account.contributions.delete.msg', 'Please delete all my contributions.');
     int userPostCount = await Groups().getUserPostCount();
@@ -876,6 +881,63 @@ class _SettingsPersonalInfoContentWidgetState extends State<SettingsPersonalInfo
     await Auth2().deleteUser();
   }
 
+  Widget _buildNamePronouncement(){
+    return Container(
+        child: Row(
+          children: [
+            Container(),//TBD
+            Visibility(visible: !_hasStoredPronouncement, child:
+              Expanded(
+                  child: GestureDetector(onTap:  _onRecordNamePronouncement, child:
+                    Text( Localization().getStringEx("", "Add name pronunciation and how you prefer to be addressed by students (Ex: Please call me Dr. Last Name,First Name, or Nickname. )"),
+                      style: Styles().textStyles?.getTextStyle("widget.info.medium.underline"),
+                    ),
+                  )
+              ),
+            ),
+            Visibility(visible: _hasStoredPronouncement, child:
+              Expanded(
+                  child: GestureDetector(onTap:  _onPlayNamePronouncement, child:
+                    Text( Localization().getStringEx("", "Your name pronunciation recording"),
+                      style: Styles().textStyles?.getTextStyle("widget.info.medium.underline"),
+                    ),
+                  )
+              ),
+            ),
+            Visibility(visible: _hasStoredPronouncement, child:
+              InkWell(onTap: _onEditNamePronouncement, child:
+                Container(width: 30, height: 30, color: Styles().colors?.fillColorPrimary,),)//TBD
+            ),
+            Container(width: 12,),
+            Visibility(visible: _hasStoredPronouncement, child:
+              InkWell(onTap: _onDeleteNamePronouncement, child:
+                Container(width: 30, height: 30, color: Styles().colors?.fillColorSecondary,),)//TBD
+            )
+          ],
+        )
+    );
+  }
+
+  void _onPlayNamePronouncement(){
+    if(_storedAudioPronouncement != null) {
+      NamePronouncementPlayer.play(_storedAudioPronouncement!);
+    }
+  }
+
+  void _onRecordNamePronouncement(){
+     SoundRecorderDialog.show(context).then((_) => setStateIfMounted(() { }));
+  }
+
+  void _onEditNamePronouncement(){
+    SoundRecorderDialog.show(context, initialRecordPath: _storedAudioPronouncement);
+  }
+
+  void _onDeleteNamePronouncement(){
+    //TBD Implement
+    Storage().setStringWithName(SoundRecorderDialog.storage_key, null);
+    setStateIfMounted(() { });
+  }
+
 
   bool get _canSave{
     return _isEmailChanged || _isNameChanged;
@@ -908,6 +970,10 @@ class _SettingsPersonalInfoContentWidgetState extends State<SettingsPersonalInfo
   bool get _hasProfilePicture {
     return (_profileImageBytes != null);
   }
+
+  bool get _hasStoredPronouncement => StringUtils.isNotEmpty(_storedAudioPronouncement);
+
+  String? get _storedAudioPronouncement => Storage().getStringWithName(SoundRecorderDialog.storage_key); //TBD implement from profile. Add profile chenge listener
 
   TextStyle? get _formFieldLabelTextStyle {
     return  Styles().textStyles?.getTextStyle("widget.item.small");