+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Introduced Name Pronouncement Dialog [#3846](https://github.com/rokwire/illinois-app/issues/3846).
 ### Changed
 - Replaced textScaleFactor usage with textScaler [#3832](https://github.com/rokwire/illinois-app/issues/3832).
 - Acknowledge the latest changes for renewing mobile access [#3840](https://github.com/rokwire/illinois-app/issues/3840).